Pull out ColumnReferenceParser
