+++ dbml-parser/src/main/kotlin/com/zynger/floorplan/lex/ColumnParser.kt
@@ -2,11 +2,9 @@ package com.zynger.floorplan.lex
 
 import com.zynger.floorplan.Column
 import com.zynger.floorplan.Reference
-import com.zynger.floorplan.ReferenceOrder
 
 object ColumnParser {
     private val COLUMN_REGEX = Regex("""("\w+"|\w+)+\s+("\w+"|\w+)(\s+\[[^]]*]|)[ ]*\n""")
-    private val COLUMN_REFERENCE_REGEX = Regex("""((ref)\s*:\s*([<>-])\s*("\w+"|\w+)\.("\w+"|\w+))""")
 
     fun parseColumns(tableName: String, columnsInput: String): List<Column> {
         return COLUMN_REGEX.findAll(columnsInput).map {
@@ -17,17 +15,7 @@ object ColumnParser {
             val notNull = columnProperties.contains("not null")
             val pk = columnProperties.contains("pk")
             val increment = columnProperties.contains("increment")
-            val reference: Reference? = if (COLUMN_REFERENCE_REGEX.containsMatchIn(columnProperties)) {
-                val referenceProperties = COLUMN_REFERENCE_REGEX.find(columnProperties)!!
-                Reference(
-                    rawValue = referenceProperties.groups[0]!!.value,
-                    fromTable = tableName,
-                    fromColumn = name,
-                    referenceOrder = ReferenceOrder.fromString(referenceProperties.groups[3]!!.value),
-                    toTable = referenceProperties.groups[4]!!.value,
-                    toColumn = referenceProperties.groups[5]!!.value
-                )
-            } else null
+            val reference: Reference? = ColumnReferenceParser.parse(tableName, name, columnProperties)
 
             Column(
                 rawValue = rawValue,
